Prevent integer multiplication overflow errors by correctly setting array data-types
Previously, some arrays were configured to be numpy.int8, and others
were initialized with a data-type based off of the highest number they
were supposed to hold. However, when these arrays were operated on with
integer arrays, there was frequent overflows that would require the
next largest data-type. Since this level of optimization is largely
unnecessary this commit generalizes those data-types, or lets them be
automatically selected as the default numpy.int64.

+++ single_agent_augmented_policy_inference.py
@@ -29,7 +29,7 @@ np.set_printoptions(precision=4)
 # Grid, number of agents, obstacle, label, action, initial and goal state configuration
 
 grid_dim = [8, 8] # [num-rows, num-cols]
-grid_map = np.array(range(0,np.prod(grid_dim)), dtype=np.int8).reshape(grid_dim)
+grid_map = np.array(range(0,np.prod(grid_dim))).reshape(grid_dim)
 
 # Create a list of tuples where the tuples have length @c num_agents and represent the joint states of the agents.
 num_agents = 1
